+++ CHANGELOG.md
@@ -13,6 +13,8 @@ CHANGED:
 
 - update of parent version to 4.2
 - update of silly-collections version to 4.35
+- update of jcommons-lang version to 5
+- update of file-worker version to 5
 - tagged 'look and feel' enum and action classes as deprecated 
 
 Version 4.30