+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Change log
 ----------------------
 
-Version 4.31-SNAPSHOT
+Version 4.31
 -------------
 
 ADDED: